requirements.txt
fastapi
uvicorn[standard]
pydantic
requests
lxml
PyYAML
scrapy
twisted
httpx
redis>=5.0.0
pytest
pytest-asyncio
respx
wsproto==1.3.2
ruff
